Time: 0 ms (100%), Space: 8.3 MB (89.07%) - LeetHub

+++ 0326-power-of-three/0326-power-of-three.cpp
@@ -4,7 +4,7 @@ public:
         if(n<=0){
             return false;
         }
-        long int temp= 1;
+        double temp= 1;
         while(temp<=n){
             if(temp==n){
                 return true;
